% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{classifyByBreaks}
\alias{classifyByBreaks}
\title{Group numerical vector (enhanced cut())}
\usage{
classifyByBreaks(x, breaks, i.class.zero = FALSE, format.level = FALSE,
  ordered.factor = FALSE)
}
\arguments{
\item{x}{numieric vector}

\item{breaks}{first lower bound and all following upper bounds}

\item{i.class.zero}{indicator for a separate group vor zero values}

\item{format.level}{indicator for levels to be formatted}

\item{ordered.factor}{to be given to cut()}
}
\value{
grouped factor vector
}
\description{
Group numerical vector and allow for group with zero values as well as formating
}
\details{
utility function for better grouping numerical valures
}
\examples{
 x <- as.integer(c(rep(0, 10), rep(1, 8), rep(2, 6), rep(3, 6), rep(4, 4), 5, 5, 5, 6, 6, 7, 8, 9))
 breaks <- c(0, 1, 2, 3, 5, 9)

 y <- classifyByBreaks(x = x, breaks = breaks, format.level = FALSE)
 table(y, useNA = "always")

 y <- classifyByBreaks(x = x, breaks = breaks, format.level = TRUE)
 table(y, useNA = "always")

 y <- classifyByBreaks(x = x, breaks = breaks, i.class.zero = 1, format.level = FALSE)
 table(y, useNA = "always")

 y <- classifyByBreaks(x = x, breaks = breaks, i.class.zero = 1, format.level = TRUE)
 table(y, useNA = "always")
}
\author{
Roland Rapold
}
\keyword{NA}
