% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{classifyByQuantiles}
\alias{classifyByQuantiles}
\title{Group numerical vector (enhanced cut())}
\usage{
classifyByQuantiles(x, n.main = 10, n.supp = 4, i.class.zero = FALSE,
  format.level = FALSE, ordered.factor = FALSE)
}
\arguments{
\item{x}{numieric vector}

\item{n.main}{number of main groups}

\item{n.supp}{number of supplement groups - total size of which equals sizt of one main group}

\item{i.class.zero}{indicator for a separate group vor zero values}

\item{format.level}{indicator for levels to be formatted}

\item{ordered.factor}{to be given to cut()}
}
\value{
grouped factor vector
}
\description{
Group numerical vector and allow for group with zero values as well as formating
}
\details{
utility function for better grouping numerical valures
}
\examples{
 data(swiss)
 summary(swiss$Agriculture)
 x            <- swiss$Agriculture
 x[c(3, 6, 9, 22)] <- NA
 n.main       <- 6
 n.supp       <- 2
 i.class.zero <- 1
 i.class.zero <- 0
 format.level <- TRUE
 y <- classifyByQuantiles(x = x, n.main = n.main, n.supp = n.supp,
                          i.class.zero = i.class.zero,
                          format.level = format.level)
 str(y)
 levels(y)
 table(y, useNA = "always")
 sum(table(y, useNA = "always"))


 format.level <- FALSE
 y <- classifyByQuantiles(x = x, n.main = n.main, n.supp = n.supp,
                          i.class.zero = i.class.zero,
                          format.level = format.level)
 str(y)
 levels(y)
 table(y, useNA = "always")
}
\author{
Roland Rapold
}
\keyword{NA}
