% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compTwoCharVects}
\alias{compTwoCharVects}
\title{Compare two string vectors}
\usage{
compTwoCharVects(vec1, vec2, showElements = FALSE, ...)
}
\arguments{
\item{vec1}{vector one}

\item{vec2}{vector two}

\item{showElements}{TRUE/FALSE for showing entries}

\item{\dots}{arguments passed to further functions}
}
\value{
comparisons of vector contents with respect to equality and optionally show entries
}
\description{
Compare two string vectors
}
\note{
under continuous developement
}
\examples{
if(base::require("car")) {
   data(Mroz, package = "carData")
   str(Mroz)
   Mroz$wc <- as.character(Mroz$wc)
   Mroz$hc <- as.character(Mroz$hc)
   compTwoCharVects(Mroz$wc, Mroz$hc)
}
}
\references{
none
}
\author{
Roland Rapold
}
