% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countDistinct.R
\name{compTwoVects}
\alias{compTwoVects}
\title{Compare two columns of a data.frame}
\usage{
compTwoVects(x, y, ...)
}
\arguments{
\item{x}{vector one}

\item{y}{vector of same length as x}

\item{\dots}{arguments passed to further functions}
}
\value{
comparisons of vector contents with respect to equality and NAs
}
\description{
Compare two columns of a data.frame
}
\note{
under continuous developement
}
\examples{
if(base::require("car")) {
   data(Mroz, package="car")
   str(Mroz)
   dim(Mroz)
   table(Mroz$wc, useNA="always")
   set.seed(1234567)
   Mroz[sample(nrow(Mroz), 5), "wc"] <- NA
   Mroz$wc_new <- Mroz$wc
   Mroz[sample(nrow(Mroz), 2), "wc"]      <- NA
   Mroz[sample(nrow(Mroz), 8), "wc_new"]  <- NA
   Mroz[sample(nrow(Mroz), 17), "wc_new"] <- "yes"
   table(Mroz$wc, useNA="always")
   table(Mroz$wc_new, useNA="always")
   #
   compTwoVects(Mroz$wc, Mroz$wc_new)
}
}
\references{
none
}
\author{
Roland Rapold
}
