% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCharacterToFactor.R
\name{convertCharacterToFactor}
\alias{convertCharacterToFactor}
\title{Convert character features to factor features in data.frames}
\usage{
convertCharacterToFactor(data)
}
\arguments{
\item{data}{[\code{data.frame}]}
}
\description{
Convert character features to factor features in data.frames
}
\examples{
\dontrun{
data("mpg", package = "ggplot2")
str(mpg)
convertCharacterToFactor(mpg)
str(mpg)}
}

