% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeFactorFeatures.R
\name{countTaskClassCharacteristics}
\alias{countTaskClassCharacteristics}
\title{Count number of factor levels for all variables, returns 0 for all non-factors}
\usage{
countTaskClassCharacteristics(task, threshold = 0L, larger.threshold = TRUE)
}
\arguments{
\item{task}{[\code{Task}]}

\item{threshold}{[\code{integer(1)}] Indicates threshold which features 
should be returned}

\item{larger.threshold}{[\code{logical(1)}] If TRUE only features with 
number of levels larger-equal threshold are returned. If set to FALSE only 
smaller-equall features are returned}
}
\description{
Count number of factor levels for all variables, returns 0 for all non-factors
}

