% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defs2.R
\name{createBaselearners2}
\alias{createBaselearners2}
\title{Create Base Learners 2}
\usage{
createBaselearners2(task)
}
\arguments{
\item{task}{[\code{Task}]\cr
Used for mtry in randomForest}
}
\description{
Create Base Learners 2
}

