% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defs3.R
\name{createBaselearners3}
\alias{createBaselearners3}
\title{Create Base Learners 2}
\usage{
createBaselearners3(task)
}
\arguments{
\item{task}{[\code{Task}]\cr
Used for mtry in randomForest}
}
\description{
Create Base Learners 2
}

