% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defs4b.R
\name{createBaselearners4b}
\alias{createBaselearners4b}
\title{Create Base Learners 2}
\usage{
createBaselearners4b(task)
}
\arguments{
\item{task}{[\code{Task}]\cr
Used for mtry in randomForest}
}
\description{
Create Base Learners 2
}

