% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTcreateBoxplot.R
\name{createBoxplotMT}
\alias{createBoxplotMT}
\title{MT: Create Boxplots with scaling or cutting of y axis
æ}
\usage{
createBoxplotMT(data, metric = mmce, opp.bound = NULL, scale = NULL,
  main = NULL, plotly = TRUE)
}
\arguments{
\item{data}{dataæ}

\item{metric}{mlr-metric}

\item{opp.bound}{lower or upper bound for y axis depending if metric should 
be minimzed or maximized.}

\item{scale}{scales "log10" or "sqrt" are allowed.}

\item{main}{title}
}
\description{
MT: Create Boxplots with scaling or cutting of y axis
æ
}

