% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defs2.R
\name{createMMPS2}
\alias{createMMPS2}
\title{MT: Make Model Multiplexer Param Set Setting}
\usage{
createMMPS2(mm, task)
}
\arguments{
\item{mm}{Model Multiplexer (from MTmakeMM)}

\item{task}{Task}
}
\description{
MT: Make Model Multiplexer Param Set Setting
}

