% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defs3.R
\name{createMMPS3}
\alias{createMMPS3}
\title{MT: Make Model Multiplexer Param Set Setting}
\usage{
createMMPS3(mm, task)
}
\arguments{
\item{mm}{Model Multiplexer (from MTmakeMM)}

\item{task}{Task}
}
\description{
MT: Make Model Multiplexer Param Set Setting
}

