% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defs4.R
\name{createMMPS4}
\alias{createMMPS4}
\title{MT: Make Model Multiplexer Param Set Setting}
\usage{
createMMPS4(mm, task)
}
\arguments{
\item{mm}{Model Multiplexer (from MTmakeMM)}

\item{task}{Task}
}
\description{
MT: Make Model Multiplexer Param Set Setting
}

