% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTasksFromOML.R
\name{createTaskFromOML}
\alias{createTaskFromOML}
\title{Make Task from OpenML data set with some preprocessing}
\usage{
createTaskFromOML(id, min.perc = 0.05, perc = 0.05, na.omit = TRUE)
}
\description{
Converts OpenML data set in tasks, merging small factor levels, remove missings, 
and remove constant features. This function name the task and prompt the user 
to name the target if this is not specified in the OpenML data set.
}

