% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrmisc.R
\docType{data}
\name{d.sport}
\alias{d.sport}
\title{Sample data}
\format{\itemize{
 \item weit
 \item kugel
 \item hoch
 \item disc
 \item stab
 \item speer
 \item punkte
}}
\description{
sport measures from 15 athletes
}
\examples{
    data(d.sport)
    print(d.sport)
}
\references{
"http://stat.ethz.ch/Teaching/Datasets/WBL/sport.dat" accessed 2016-06-03
}
\keyword{data}
