\name{d.sport}
\alias{d.sport}
\docType{data}
\title{ sample data }
\description{ sport measures from 15 athletes }
\usage{
    data(d.sport)
}
%% \arguments{ }
\format{
    \itemize{
        \item{weit}{ long jump }
        \item{kugel}{ bowl }
        \item{hoch}{ high jump }
        \item{disc}{ disc }
        \item{stab}{ stick... }
        \item{speer}{ stick... }
        \item{punkte}{ points }
    }
}
%% \details{ }
%% \source{ }
\references{ Originally from: "http://stat.ethz.ch/Teaching/Datasets/WBL/sport.dat" accessed 2016-06-03 }
\examples{
    data(d.sport)
    print(d.sport)
}
\keyword{data}
