% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatPValue.R
\name{encodeUTF8}
\alias{encodeUTF8}
\title{Encode text to UTF-8}
\usage{
encodeUTF8(df, ...)
}
\arguments{
\item{df}{data.frame or character vector}

\item{\dots}{arguments passed to further functions}
}
\value{
formatted object
}
\description{
Encode text to UTF-8
}
\details{
utility function for formating
}
\note{
maybe of most use under Windows-Environment
}
\examples{
d.test <- data.frame(a = 1:10,
                     b = letters[1:10],
                     c = c("ä", "ö", "ü", "è", "à", "\%", "&", "¢", "@", "¬"))
d.test
encodeUTF8(d.test)
}
\references{
none
}
\seealso{
other utility-functions in this R-package
}
\author{
Roland Rapold
}
