% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSubSubList.R
\name{extractSubSubList}
\alias{extractSubSubList}
\title{Extract list from list}
\usage{
extractSubSubList(xs, element1, element2)
}
\arguments{
\item{xs}{list}

\item{element1}{character from upper list}

\item{element2}{character from inner list}
}
\description{
Extract list from list
}

