% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatifiedBinaryData.R
\name{getStratifiedBinaryData}
\alias{getStratifiedBinaryData}
\title{Get stratified samples from data with binary target}
\usage{
getStratifiedBinaryData(task, size = 20L)
}
\arguments{
\item{task}{[\code{Task}]}

\item{size}{[\code{integer(1)}] Total size of sample.}
}
\description{
Get stratified samples from data with binary target
}

