% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepFunctions.R
\name{grepColFactors}
\alias{grepColFactors}
\title{Get column names of data.frame with factors}
\usage{
grepColFactors(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
grepColFactors() returns variables are returned and indicated if also defined as
'ordered'.
}
\description{
Of a data.frame, get the attributes with class 'factor' and 'ordered factor'.
}
\details{
Factor variables are returned and indicated if also defined as 'ordered'.
}
\note{
under continuous developement
}
\examples{
if(require("car"))
{
  data(Mroz, package="car")
  Mroz$wc_ord <- as.ordered(Mroz$wc)
  print(grepColFactors(Mroz))
}
}
\references{
none
}
\seealso{
\code{\link{grepColEntries}} \code{\link{grepColNames}} \code{\link{grepColNA}}
\code{\link{grepObjNames}} \code{\link{grepNotCompleteEntries}} \code{\link{grepMultipleEntries}}
\code{\link{grepRowNames}} \code{\link{grepColNegNum}}
}
\author{
Roland Rapold
}
