% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepFunctions.R
\name{grepColNA}
\alias{grepColNA}
\title{Get attributes of data.frame which contain NAs}
\usage{
grepColNA(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{\dots}{Arguments passed to further functions}
}
\value{
Column names which values contain NAs or indication about NAs.
}
\description{
Get attributes of data.frame which contain NAs.
}
\details{
If the input is a vector, grepColNA() indicates if NAs are present. If the input is a
data.frame, grepColNA() returns the colums which contain NAs.
}
\note{
under continuous developement
}
\examples{
if(require("MASS"))
{
    # load example data
    data(crabs, package="MASS")
    head(crabs)
    print(grepColNA(crabs))
    tt <- crabs
    tt[122,2] <- NA
    tt[22, 4] <- NA
    print(grepColNA(tt))
    print(grepColNA(c(pi, NA, 4.000000, 5.000000, 5, 6, 5, 6)))
}
}
\references{
none
}
\seealso{
grepColEntries grepColNames grepObjNames grepNotCompleteEntries grepMultipleEntries
grepRowNames grepColNegNum grepColFactors
}
\author{
Roland Rapold
}
