% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepFunctions.R
\name{grepColNames}
\alias{grepColNames}
\title{Get column names of data.frame matching a search string}
\usage{
grepColNames(d.frame, part.name, sort.col = "nr", ...)
}
\arguments{
\item{d.frame}{data.frame}

\item{part.name}{Search string in column names}

\item{sort.col}{specify output column to sort by}

\item{\dots}{Arguments passed to further functions}
}
\value{
grepColNames() returns the found column names and details
}
\description{
Specify a character string and find the column names in a data.frame containing this
character string.
}
\details{
Unlike using '==' or '%in%', here search is performed using grep() in a fashion of
'like', meaning substrings are recognized.
}
\note{
under continuous developement
}
\examples{
if(require("MASS"))
{
    data(crabs, package="MASS")
    head(crabs)
    print(grepColNames(crabs, "W"))
    print(grepColNames(crabs, "e"))
}
}
\references{
none
}
\seealso{
grepColEntries grepColNA grepObjNames grepNotCompleteEntries grepMultipleEntries grepRowNames grepColNegNum grepColFactors
}
\author{
Roland Rapold
}
