% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepFunctions.R
\name{grepColNegNum}
\alias{grepColNegNum}
\title{Get column names of data.frame who contain negative values}
\usage{
grepColNegNum(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
grepColNegNum() returns the column number and names of columns containing negabive
numeric values.
}
\description{
grepColNegNum() returns the column numbers and names of the numerical columns
(numeric and integer) which contain negative numbers
}
\details{
Only numeric attributes (numeric and integer) are searched and reported.
}
\note{
under continuous developement
}
\examples{
if(require("MASS"))
{
    require("car")
    data(Mroz)
    head(Mroz)
    x <- Mroz
    print(grepColNegNum(Mroz))
}
}
\references{
none
}
\seealso{
\code{\link{grepColEntries}} \code{\link{grepColNames}} \code{\link{grepColNA}}
\code{\link{grepObjNames}} \code{\link{grepNotCompleteEntries}} \code{\link{grepMultipleEntries}}
\code{\link{grepRowNames}} \code{\link{grepColFactors}}
}
\author{
Roland Rapold
}
