% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepFunctions.R
\name{grepMultipleEntries}
\alias{grepMultipleEntries}
\title{Records with duplicates in selected columns}
\usage{
grepMultipleEntries(d.frame, var, ...)
}
\arguments{
\item{d.frame}{data.frame}

\item{var}{variable(s) to search for duplicates}

\item{\dots}{Arguments passed to further functions}
}
\value{
Subset of data.frame for which selection conditions are met
}
\description{
Subset a data.frame wich multiple entries in one or more columns.
}
\details{
The multiple entries are found with duplicated() and the whole records of these entries
are returned.
}
\note{
under continuous developement
}
\examples{
if(require("MASS"))
{
    data(crabs, package="MASS")
    #
    # - with one column
    print(head(grepMultipleEntries(crabs, "FL"), n=10))
    #
    # - with two columns
    print(head(grepMultipleEntries(crabs, c("sex", "FL"))))
}
}
\references{
none
}
\seealso{
grepColEntries grepColNA grepObjNames grepNotCompleteEntries grepColNames grepRowNames grepColNegNum grepColFactors
}
\author{
Roland Rapold
}
