% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{grepNotCompleteEntries}
\alias{grepNotCompleteEntries}
\title{Get incomplete records in a data.frame}
\usage{
grepNotCompleteEntries(d.frame, ...)
}
\arguments{
\item{d.frame}{data.frame}

\item{\dots}{Arguments passed to further functions}
}
\value{
Subset of data.frame where records contain NAs.
}
\description{
grepNotCompleteEntries returns the records of a data.frame containing NAs.
}
\details{
The input is a data.frame of which the records are returned that contain NAs in any
values.
}
\note{
under continuous developement
}
\examples{
if(require("MASS"))
{
    data(crabs, package="MASS")
    df <- crabs
    df[33, 4] <- NA
    df[55, 5] <- NA
    df[77, 6] <- NA
    df[99, 7] <- NA
    print(grepNotCompleteEntries(df))
}
}
\references{
none
}
\seealso{
\code{\link{grepColEntries}} \code{\link{grepColNames}} \code{\link{grepColNA}}
\code{\link{grepObjNames}} \code{\link{grepMultipleEntries}} \code{\link{grepRowNames}}
\code{\link{grepColNegNum}} \code{\link{grepColFactors}}
}
\author{
Roland Rapold
}
