% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepFunctions.R
\name{grepObjNames}
\alias{grepObjNames}
\title{Find objects whose names contain search string}
\usage{
grepObjNames(part.name, ...)
}
\arguments{
\item{part.name}{search string in object names}

\item{\dots}{Arguments passed to further functions}
}
\value{
column names which values contain NAs or indication about NAs
}
\description{
The function grepObjNames() gets the objects whose names contain the specified
search string.
}
\details{
Unlike subset which works with '==' or '\%in\%', here search is performed in a fashion of
'like', meaning substrings are recognized.
}
\note{
under continuous developement
}
\examples{
    a  <- 1
    aa <- 1
    b  <- 1
    ab <- 1
    A  <- 1
    Aa <- 1
    b  <- 1
    print(grepObjNames("a"))
    print(grepObjNames("A"))
    print(grepObjNames("b"))
}
\references{
none
}
\seealso{
\code{\link{grepColEntries}} \code{\link{grepColNames}} \code{\link{grepColNA}}
\code{\link{grepNotCompleteEntries}} \code{\link{grepMultipleEntries}} \code{\link{grepRowNames}}
\code{\link{grepColNegNum}} \code{\link{grepColFactors}}
}
\author{
Roland Rapold
}
