% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepFunctions.R
\name{grepRowNames}
\alias{grepRowNames}
\title{Get row names of data.frame matching a search string}
\usage{
grepRowNames(d.frame, part.name, ...)
}
\arguments{
\item{d.frame}{data.frame}

\item{part.name}{Search string in row names}

\item{\dots}{Arguments passed to further functions}
}
\value{
Number and names of found row names.
}
\description{
Specify a character string and find the row names in a data.frame containing this
character string.
}
\details{
Unlike using '==' or '%in%', here search is performed using grep() in a fashion of
'like', meaning substrings are recognized.
}
\note{
under continuous developement
}
\examples{
if(require("MASS"))
{
    data(crabs, package="MASS")
    head(crabs)
    rownames(crabs) <- paste0("row_", rownames(crabs))
    print(grepRowNames(crabs, "16"))
}
}
\references{
none
}
\seealso{
grepColEntries grepColNA grepObjNames grepNotCompleteEntries grepMultipleEntries grepColNames grepColNegNum grepColFactors
}
\author{
Roland Rapold
}
