% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listhead.R
\name{listhead}
\alias{listhead}
\title{Head of data.fames in list}
\usage{
listhead(x, ...)

listhead(x, ...)
}
\arguments{
\item{x}{list}

\item{...}{further head args}

\item{x}{list}

\item{...}{further tail args}
}
\description{
Head of data.fames in list

Tail of data.fames in list
}

