% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{listObjContrasts}
\alias{listObjContrasts}
\title{Get contrasts of data.frame}
\usage{
listObjContrasts(d.frame, ...)
}
\arguments{
\item{d.frame}{data.frame}

\item{\dots}{arguments passed to further functions}
}
\value{
contrasts of factor variables of data.frame
}
\description{
Get the contrasts of all factor variables in a data.frame.
}
\details{
Get the contrasts of all factor variables in a data.frame.
}
\note{
under continuous developement
}
\examples{
if(require("MASS")){
    data(crabs, package="MASS")
    print(listObjContrasts(crabs))
}
}
\references{
none
}
\seealso{
other utility-functions in this R-package
}
\author{
Roland Rapold
}
