% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objDescr.R
\name{listObjSizes}
\alias{listObjContrasts}
\alias{listObjSizes}
\title{get properties of objects easily.}
\usage{
listObjSizes(...)
}
\arguments{
\item{\dots}{arguments passed to further functions}
}
\value{
sizes and contrasts of loaded objects
}
\description{
get sizes and contrasts of objects.
}
\details{
function for the description of properties of objects rather than processing
of data.
}
\note{
under continuous developement
}
\examples{

if(require("MASS")){
    # Example - loading data
    data(crabs, package="MASS")
    print(head(crabs))
    #
    # Example for 'listObjContrasts
    print(listObjContrasts(crabs))
    #
    # Example for 'listObjSizes'
    print(listObjSizes())
}

}
\author{
Roland Rapold
}
\references{
none
}
\seealso{
other utility-functions in this R-package
}

