% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTaddUniqueName.R
\name{MTaddUniqueName}
\alias{MTaddUniqueName}
\title{Master's Thesis: Add unique name to experiment result}
\usage{
MTaddUniqueName(data)
}
\arguments{
\item{data}{data}
}
\description{
Master's Thesis: Add unique name to experiment result
}

