% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTdefs.R
\name{MTcreateBaselearners}
\alias{MTcreateBaselearners}
\title{Create Base Learners}
\usage{
MTcreateBaselearners(task)
}
\arguments{
\item{task}{[\code{Task}]\cr
Used for mtry in randomForest}
}
\description{
Create Base Learners
}

