% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTdefs.R
\name{MTcreateMMPS}
\alias{MTcreateMMPS}
\title{MT: Make Model Multiplexer Param Set Setting}
\usage{
MTcreateMMPS(mm, task)
}
\arguments{
\item{mm}{Model Multiplexer (from MTmakeMM)}

\item{task}{Task}
}
\description{
MT: Make Model Multiplexer Param Set Setting
}

