% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractClassifTaskInfo.R
\name{MTextractClassifTaskInfo}
\alias{MTextractClassifTaskInfo}
\title{Extract ClassifTask Info}
\usage{
MTextractClassifTaskInfo(task)
}
\description{
Extract ClassifTask Info
}
\examples{
\dontrun{
library(mlr)
library(xtable)
tasks = list(pid.task, pid.task)
x = lapply(tasks, function(x) extractTaskInfo(x))
x = data.frame(Reduce(rbind, x))
xtable(x)
}
}

