% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteSuppressNA.R
\name{pasteSuppressNA}
\alias{pasteSuppressNA}
\title{Paste function suppressing NAs}
\usage{
pasteSuppressNA(..., sep = "")
}
\arguments{
\item{...}{[\code{numeric/factor columns(n)}] Values to paste}

\item{sep}{[\code{character(1)}] Seperator}
}
\description{
Source http://stackoverflow.com/questions/13673894/suppress-nas-in-paste (minor adjustments)
}

