% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pValues.R
\name{pFromT}
\alias{pFromT}
\title{Calculate p-values from t-statistics.}
\usage{
pFromT(t_value = 1.96, df = 10, two.sided = FALSE)
}
\arguments{
\item{t_value}{The t-statistics to be converted into a p-Value.}

\item{df}{Degree of freedom in t-statistics.}

\item{two.sided}{Indicator of the statistics be calculatede as two-sided (TURE) or one-sided (FALSE).}
}
\value{
corresponding p-Value from the t-, z- or F-statistics
}
\description{
Calculate p-values from t-statistics.
}
\examples{
        pFromT(t_value=1.96, df=10)
        pFromT(t_value=1.96, df=100)
        pFromT(t_value=1.96, df=1000)
        pFromT(t_value=1.96, df=1000, two.sided = TRUE)

        x <- c(seq(0.0001, 0.6, 0.01), NA, seq(0.6001, 1.0, 0.01))
        p_value <- pFromT(t_value=x, df=100)
        p_value
        formatPValue(p_value, digits=3)

}
\seealso{
\code{Link{pFromZ}} and \code{Link{pFromF}}
}
\author{
Roland Rapold
}
\keyword{p-value}
