% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pValues.R
\name{pFromZ}
\alias{pFromZ}
\title{Calculate p-values from z-statistics.}
\usage{
pFromZ(z_value = 1.96, two.sided = FALSE)
}
\arguments{
\item{z_value}{The z-statistics to be converted into a p-Value.}

\item{two.sided}{Indicator of the statistics be calculatede as two-sided (TURE) or one-sided (FALSE).}
}
\value{
corresponding p-Value from the t-, z- or F-statistics
}
\description{
Calculate p-values from z-statistics.
}
\details{
utility function for conversion of p-values
}
\examples{
        pFromF(f_value=35.68, df1=2,  df2=5)
        pFromF(f_value=35.68, df1=2,  df2=5, two.sided=TRUE)
        pFromF(f_value= 1.96, df1=10, df2=10)
        pFromF(f_value= 1.96, df1=10, df2=100)
        pFromF(f_value= 1.96, df1=10, df2=1000)
        pFromF(f_value= 1.96, df1=10, df2=1000, two.sided=TRUE)
        pFromF(f_value= 1.96, df1=10, df2=1000, two.sided=TRUE)
}
\seealso{
\code{Link{pFromT}} and \code{Link{pFromF}}
}
\author{
Roland Rapold
}
\keyword{p-value}
