% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{profileData}
\alias{profileData}
\title{Run shiny-appt to profile data and create report}
\usage{
profileData(...)
}
\arguments{
\item{\dots}{other arguments}
}
\value{
shiny app and reports in different formats
}
\description{
Run shiny-appt to profile data and create report
}
\note{
under continuous developement
}
\references{
none
}
\author{
Roland Rapold
}
