% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeFactorFeatures.R
\name{removeFactorFeatures}
\alias{removeFactorFeatures}
\title{Remove Features from Task which have too man factor levels}
\usage{
removeFactorFeatures(task, n.levels = 32L)
}
\arguments{
\item{task}{[\code{Task}] from mlr.}

\item{n.levels}{[\code{integer(1)}] threshold. Default is 32, which refers
to the maximal amount of levels allowed in \code{randomForest}.}
}
\description{
Remove Features from Task which have too man factor levels
}

