\name{rrMisc-package}
\alias{rrMisc-package}
\alias{rrMisc}
\docType{package}
\title{some R-helpers}
\description{some helpers of R to process data smoothly}
\details{
\tabular{ll}{
    Package: \tab rrMisc\cr
    Type: \tab Package\cr
    Version: \tab 0.13\cr
    Date: \tab 2016-06-12\cr
    License: \tab --
}
Some rather small helpers like the 'grep' funktions to facilitate handling of objects.
}
\author{ Author and Maintainer: roland.rapold@helsana.ch }
\references{ }
\keyword{ package }
\seealso{ other functions in this R-packge }
\examples{
}
