\name{sample_data_sport}
\alias{sample_data_sport}
\docType{data}
\title{ sample data }
\description{ sample data }
\usage{
    str(sample_data_sport)
}
\arguments{
    sport measures from 15 athletes
    \item{weit}{ long jump }
    \item{kugel}{ bowl }
    \item{hoch}{ high jump }
    \item{disc}{ disc }
    \item{stab}{ stick... }
    \item{speer}{ stick... }
    \item{punkte}{ points }
}
%% \format{ }
%% \details{ }
%% \source{ }
%% \references{ }
%% \examples{
%%     data(sample_data_sport)
%%     print(sample_data_sport)
%% }
%% \keyword{graphics}
