% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{tableCountPart}
\alias{tableCountPart}
\title{Generate table with counts, parts and cumulative parts}
\usage{
tableCountPart(x, varName = "var", anzName = "anz", antName = "ant",
  csumName = FALSE, ordCol = 1, ordDir = "asc", rowLimit = FALSE,
  ...)
}
\arguments{
\item{x}{data.frame}

\item{varName}{name of attribute to list}

\item{anzName}{column name for counts}

\item{antName}{column name for parts}

\item{csumName}{column name for cumulative parts}

\item{ordCol}{number of column to sort}

\item{ordDir}{sort order ('asc' for ascending or 'desc' for descending)}

\item{rowLimit}{number of rows with detailed counts and parts}

\item{\dots}{arguments passed to further functions}
}
\value{
data.frame with computed statistics
}
\description{
Generate table with counts, parts and cumulative parts. Allow to cut table after
predefined length
}
\note{
under continuous developement
}
\examples{
   data(mtcars, package = "datasets")
   str(mtcars)
   table(mtcars$gear)
   tableCountPart(x = mtcars$gear
           , varName = "Gänge"
           , anzName = "Anzahl"
           , antName = "Anteil"
           , csumName = "Anteil kumuliert"
           , ordCol = 1
           , ordDir = "desc"
           , rowLimit = 20
           )
   tableCountPart(x = mtcars$carb
           , varName = "Vergaser"
           , anzName = "Anzahl"
           , antName = "Anteil"
           , csumName = "Anteil kumuliert"
           , ordCol = 1
           , ordDir = "asc"
           , rowLimit = 20
           )
   tableCountPart(x = mtcars$carb
           , varName = "Vergaser"
           , anzName = "Anzahl"
           , antName = "Anteil"
           , csumName = "Anteil kumuliert"
           , ordCol = 1
           , ordDir = "asc"
           , rowLimit = 4
           )
   tableCountPart(x = mtcars$carb
           , varName = "Vergaser"
           , anzName = "Anzahl"
           , antName = "Anteil"
           )
}
\references{
none
}
\author{
Roland Rapold
}
