% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{tableCountPart_2d}
\alias{tableCountPart_2d}
\title{Generate table with counts an parts in various different representations.}
\usage{
tableCountPart_2d(x = 1, y = 1, z = 1, precDigit = 1,
  x_prefix = "", y_prefix = "", debug = 0, big.mark = FALSE, ...)
}
\arguments{
\item{x}{data vector together with y as alternative to z}

\item{y}{data vector together with x as alternative to z}

\item{z}{data.frame as alternative to x and y}

\item{precDigit}{number of digits for result}

\item{x_prefix}{prefix for label in x-axis}

\item{y_prefix}{prefix for label in y-axis}

\item{debug}{not relevant}

\item{big.mark}{= thousand separator}

\item{\dots}{arguments passed to further functions}
}
\value{
data.frame with computed statistics
}
\description{
Generate table with counts and parts in various different representations. The
differen representations are also du to marginal totals in the different directions.
}
\note{
under continuous developement
}
\examples{
   str(mtcars)
   table(mtcars$gear)
   table(mtcars$carb)
   precDigit  <- 1

   x <- mtcars$gear               # Auspraegungen in Resultat in Reihen
   y <- mtcars$carb               # Auspraegungen in Resultat in Spalten
   z <- as.data.frame.matrix(table(x, y))

   tableCountPart_2d(x = x, y = y, precDigit = 3) [[4]]
   tableCountPart_2d(z = z, precDigit = 3)[[4]]

   tableCountPart_2d(z = z, precDigit = 1)[[5]]
   tableCountPart_2d(z = z, precDigit = 1)[[6]]
   tableCountPart_2d(z = z, precDigit = 1)[[7]]
   tableCountPart_2d(z = z, precDigit = 0)$abs
   tableCountPart_2d(z = z, precDigit = 0)$abs_ant_col
   tableCountPart_2d(z = z, precDigit = 1)
}
\references{
none
}
\author{
Roland Rapold
}
