% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{testGranularity}
\alias{testGranularity}
\title{test granularity of attributes in data.frame}
\usage{
testGranularity(d.data, var, verbose = NULL, ...)
}
\arguments{
\item{d.data, }{data.frame}

\item{var, }{vector of attributes in d.data}

\item{verbose, }{add additional output}

\item{\dots}{arguments passed to further functions}
}
\value{
output describing the granularity
}
\description{
Compare granularity of whole data.frame and of provided attributes
}
\note{
under continuous developement
}
\examples{
cat("\\nBeispieldaten\\n")
d.data <- data.frame(a = 1:20,
                     b = rep(1:10, 2),
                     c = letters[1:20],
                     d = rep(letters[1:10], 2))

cat("\\nBetrachtung eines Attributes\\n")
var <- c("a")
testGranularity(d.data = d.data, var = var)
testGranularity(d.data = d.data, var = var, verbose = FALSE)

cat("\\nBetrachtung eines Attributes\\n")
var <- c("b")
testGranularity(d.data = d.data, var = var)

cat("\\nBetrachtung von zwei Attributen\\n")
var <- c("a", "b")
testGranularity(d.data = d.data, var = var)

cat("\\nBetrachtung von drei Attributen\\n")
var <- c("a", "b", "c", "d")
testGranularity(d.data = d.data, var = var)

cat("\\nBetrachtung mit einem Wert NA\\n")
d.data[16, 3] <- NA
var <- c("a", "b", "c", "d")
testGranularity(d.data = d.data, var = var)

cat("\\nBeispiel ChickWeight mit einem Attribut\\n")
var <- c("Chick")
testGranularity(d.data = ChickWeight, var = var)

cat("\\nBeispiel ChickWeight mit zwei Attributen\\n")
var <- c("Chick", "Diet")
testGranularity(d.data = ChickWeight, var = var)

cat("\\nBeispiel ChickWeight mit drei Attributen\\n")
var <- c("Chick", "Diet", "Time")
testGranularity(d.data = ChickWeight, var = var)
}
\references{
none
}
\author{
Roland Rapold
}
