% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{univarModDataCont}
\alias{univarModDataCont}
\title{Graph dependent variable vs. list of independent variables}
\usage{
univarModDataCont(d_data, var_dep, var_indep, label_dep, label_indep, ...)
}
\arguments{
\item{d_data}{data.frame}

\item{var_dep}{name of dependent variable}

\item{var_indep}{name of independent variable(s)}

\item{label_dep}{label of dependent variable}

\item{label_indep}{label of independent variable(s)}

\item{\dots}{arguments passed to further functions}
}
\value{
data-frame with percentage as number and character string
}
\description{
Graph dependent variable vs. list of independent variables
}
\details{
Creates xyplots for all combinations
}
\note{
under continuous developement
}
\examples{
 data(mtcars, package = "datasets")
 str(mtcars)
 univarModDataCont(d_data = mtcars
                   , var_dep = "mpg"
                   , label_dep = "Miles per Gallon"
                   , var_indep = c("cyl", "disp", "hp", "wt")
                   , label_indep = c("Zylinder", "disp", "PS", "Gewicht"))
 univarModDataCont(d_data = mtcars
                   , var_dep = "mpg"
                   , label_dep = "Miles per Gallon"
                   , var_indep = c("cyl", "hp")
                   , label_indep = c("Zylinder", "PS"))
}
\references{
none
}
\seealso{
see other functions in this R-package
}
\author{
Roland Rapold
}
