% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{univarModDataFact}
\alias{univarModDataFact}
\title{Graph influence of factors on numeric variable}
\usage{
univarModDataFact(
  d_data,
  var_dep_n,
  var_indep_f,
  label_dep_n,
  label_indep_f,
  split = TRUE,
  split_nr = 4,
  ylimits = NULL,
  plot_dims = c(15, 20),
  ...
)
}
\arguments{
\item{d_data}{data.frame}

\item{var_dep_n}{name of dependent variable (numeric)}

\item{var_indep_f}{name of independent variable(s) (factor)}

\item{label_dep_n}{label of dependent variable}

\item{label_indep_f}{label of independent variable(s)}

\item{split}{allow for split into new graphics}

\item{split_nr}{upper limit of factors per graphic}

\item{ylimits}{y limits}

\item{plot_dims}{plot dimensions}

\item{\dots}{arguments passed to further functions}
}
\value{
data-frame with percentage as number and character string
}
\description{
Graph influence of factors on numeric variable
}
\details{
Creates plot.design for all combinations
}
\note{
under continuous developement
}
\examples{
 data(npk, package = "datasets")
 str(npk)
 univarModDataFact(d_data = npk,
                   var_dep_n = "yield",
                   var_indep_f = c("N", "P", "K", "block"),
                   label_dep_n = "yield",
                   label_indep_f = c("N", "P", "K", "block"))

 data(Mroz, package = "carData")
 str(Mroz)
 univarModDataFact(d_data = Mroz,
                   var_dep_n = "age",
                   var_indep_f = c("wc", "hc"),
                   label_dep_n = "Alter",
                   label_indep_f = c("WC", "HC"))
 # , split=TRUE
 # , split_nr=9, ylim=c(0, 0.55)
 # , plot_dims=c(28, 24))
}
\references{
none
}
\seealso{
see other functions in this R-package
}
\author{
Roland Rapold
}
