% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countDistinct.R
\name{countDistinct}
\alias{countDistinct}
\title{SQL-command 'count(distinct ...)' convreted to R}
\usage{
countDistinct(x, ...)
}
\arguments{
\item{x}{Vector or data.frame in which the number of different entries are counted.}

\item{\dots}{arguments passed to further functions}
}
\value{
named vector of numbers

- count:  number of different entries (without NAs)

- NAs:    index of NAs present (0=no, 1=yes)
}
\description{
SQL-command 'count(distinct ...)' convreted to R
}
\note{
under continuous developement
}
\examples{

print("Example for vector")
                      c(pi, NA, 4.000000, 5.000000, 5, 6, 5, 6)
               length(c(pi, NA, 4.000000, 5.000000, 5, 6, 5, 6))
        countDistinct(c(pi, NA, 4.000000, 5.000000, 5, 6, 5, 6))
       length(na.omit(c(pi, NA, 4.000000, 5.000000, 5, 6, 5, 6)))
countDistinct(na.omit(c(pi, NA, 4.000000, 5.000000, 5, 6, 5, 6)))
#
if(require("MASS"))
{
    print("Example for data.frame")
    data(crabs, package="MASS")
    str(crabs)
    print(countDistinct(crabs[, c("sp")]))
    print(countDistinct(crabs[, c("sex")]))
    print(countDistinct(crabs[, c("FL")]))
}

}
\references{
none
}
\author{
Roland Rapold
}
