% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatPValue.R
\name{getContStat}
\alias{getContStat}
\title{Get statistics of contingency table}
\usage{
getContStat(d.crosstab, switchPosNeg = FALSE)
}
\arguments{
\item{d.crosstab}{= contincency table in class 'table}

\item{switchPosNeg}{= switching positive and negative classes}

\item{\dots}{arguments passed to further functions}
}
\value{
statistics of contingency table
}
\description{
Get statistics of contingency table
}
\details{
utility function for analyzing contingency tables
}
\note{
to be refined
}
\examples{
d.cont <- matrix(c(65.64, 1.95, 27.81, 4.6), ncol = 2, byrow = TRUE)
colnames(d.cont) <- c("true 0", "true 1")
rownames(d.cont) <- c("fit 0", "fit 1")
d.cont <- as.table(d.cont)
getContStat(d.cont)
}
\references{
none
}
\seealso{
other utility-functions in this R-package
}
\author{
Roland Rapold
}
