% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{grepColEntries}
\alias{grepColEntries}
\title{Subset of data.frame containing string}
\usage{
grepColEntries(d.frame, col.name, part.name, ...)
}
\arguments{
\item{d.frame}{data.frame}

\item{col.name}{Name of column}

\item{part.name}{Search string in column names}

\item{\dots}{Arguments passed to further functions}
}
\value{
Subset of data.frame for which selection conditions are met.
}
\description{
Get a subset of a data.frame with specified string in a given column.
}
\details{
Unlike subset which works with '==' or '\%in\%', here search is performed using grep() in
a fashion of 'like', meaning substrings are recognized.
}
\note{
under continuous developement
}
\examples{
if(require("MASS"))
{
    data(crabs, package="MASS")
    print(head(grepColEntries(crabs, "sex", "M")))
    print(head(grepColEntries(crabs, "index", "2")))
}
}
\references{
none
}
\seealso{
\code{\link{grepColNames}} \code{\link{grepColNA}} \code{\link{grepObjNames}}
\code{\link{grepNotCompleteEntries}} \code{\link{grepMultipleEntries}} \code{\link{grepRowNames}}
\code{\link{grepColNegNum}} \code{\link{grepColFactors}}
}
\author{
Roland Rapold
}
