% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepFunctions.R
\name{grepColEntries}
\alias{grepColEntries}
\alias{grepColFactors}
\alias{grepColNA}
\alias{grepColNames}
\alias{grepColNegNum}
\alias{grepMultipleEntries}
\alias{grepNotCompleteEntries}
\alias{grepObjNames}
\alias{grepRowNames}
\title{obtain grep components of object easily}
\usage{
grepColEntries(d.frame, col.name, part.name, ...)
}
\arguments{
\item{d.frame}{R-object}

\item{col.name}{Name of column.}

\item{part.name}{Search string in column names.}

\item{\dots}{Arguments passed to further functions.}
}
\value{
Column names, values containing search string or entries not complete or multiple entries.
}
\description{
obtain grep components of object easily
}
\details{
functions facilitating handling of objects
}
\note{
under continuous developement
}
\author{
Roland Rapold
}
\references{
none
}
\seealso{
other utility-functions in this R-package
}

