% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepFunctions.R
\name{grepColNegNum}
\alias{grepColNegNum}
\title{Get column names of data.frame who contain negative values}
\usage{
grepColNegNum(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
grepColNegNum() returns the column number and names of columns containing negabive
numeric values.
}
\description{
grepColNegNum() returns the column numbers and names of the numerical columns
(numeric and integer) which contain negative numbers
}
\details{
Only numeric attributes (numeric and integer) are searched and reported.
}
\note{
under continuous developement
}
\examples{
if(require("MASS"))
{
    data(Mroz, package="car")
    head(Mroz)
    print(grepColNegNum(Mroz))
}
}
\references{
none
}
\seealso{
grepColEntries grepColNA grepObjNames grepNotCompleteEntries grepMultipleEntries
grepRowNames grepColNames grepColFactors
}
\author{
Roland Rapold
}
