% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{listObjSizes}
\alias{listObjSizes}
\title{Show sizes of R-objects in Memory}
\usage{
listObjSizes(...)
}
\arguments{
\item{\dots}{arguments passed to further functions}
}
\value{
List of sizes of R-objects in memory.
}
\description{
Show sizes of R-objects in Memory
}
\details{
listObjSizes() returns a list of all objects in memory and lists them in order of
increasing size. Object size is reported by the function utils::object.size()
}
\note{
under continuous developement
}
\examples{
if(require("MASS")){
    data(crabs, package="MASS")
    print(listObjSizes())
}
}
\references{
none
}
\seealso{
other utility-functions in this R-package
}
\author{
Roland Rapold
}
