% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrmisc.R
\name{rrMisc-package}
\alias{rrMisc-package}
\title{Some R-helpers}
\description{
Some helpers of R to process data smoothly
}
\details{
The main function are the 'grep' functions for quickly accessing information of data objects and
the descrTable() function to generate descriptive statistics of a data object.
}
