\name{rrMisc-package}
\alias{rrMisc-package}
\alias{rrMisc}
\docType{package}
\title{some R-helpers}
\description{some helpers of R to process data smoothly}
\details{
\tabular{ll}{
    Package: \tab rrMisc\cr
    Type: \tab Package\cr
    Version: \tab 0.16\cr
    Date: \tab 2016-07-28\cr
    License: \tab GPL (>=2)
}
Some rather small helpers like the 'grep' funktions to facilitate handling of objects.
}
\author{ Author and Maintainer: roland.rapold@alumni.ethz.ch }
%% \references{ }
\keyword{ package }
\seealso{ other functions in this R-packge }
%% \examples{ }
