% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countDistinct.R
\name{testGranularity}
\alias{testGranularity}
\title{test granularity of attributes in data.frame}
\usage{
testGranularity(d.data, var, ...)
}
\arguments{
\item{d.data, }{data.frame}

\item{var, }{vector of attributes in d.data}

\item{\dots}{arguments passed to further functions}
}
\value{
output describing the granularity
}
\description{
Compare granularity of whole data.frame and of provided attributes
}
\note{
under continuous developement
}
\examples{
d.data <- data.frame(a = 1:20,
                     b = rep(1:10, 2),
                     c = letters[1:20],
                     d = rep(letters[1:10], 2))
var <- c("a")
testGranularity(d.data = d.data, var = var)

var <- c("b")
testGranularity(d.data = d.data, var = var)

var <- c("a", "b")
testGranularity(d.data = d.data, var = var)

var <- c("a", "b", "c", "d")
testGranularity(d.data = d.data, var = var)
}
\references{
none
}
\author{
Roland Rapold
}
